Switch to jupnp
Jupnp Jetty implementation has some serious flaws, so we needed to override a lot of logic ourselves, and apply some nasty hacks.

This also means that we have to switch to a more recent Jetty. We use 12 now, which means we have to patch jupnps client logic.

By the way does this also mean we have to drop sparkjava, which seams to be an abandoned project now. We will re-implement using jetty directly.

By the way did this also mean we upgrade to Java 21.