BRAKING CHANGE: Migration from Cling to jupnp, and upgrade to Java 21
MAIN CHANGES

* Switch to jupnp from cling as DLNA implementation
* Droped Sparkframework for content delivery, switched to Jetty Servlets
* Upgraded Java to 21 from 11

REQUIRED CONFIGURATION CHANGES

* PUBLIC_HTTP_PORT (defaulting to 9301) is now used for content delivery as well as for DLNA requests
* PUBLIC_BASE_URL is no longer mandatory, even if using ENABLE_PREFETCHING

The PUBLIC_HTTP_PORT is now always used, so you need to configure it if our default of 9301 is already occupied. Ephemeral port binding is no longer supported.

SOME BACKGROUND

The Cling Project is no longer maintained. Jupnp has forked the code base and it's the most active fork, so we switched to that.

The Jetty adapters for Jupnp don't seem to work because of some major defect. We had to re-implement some logic here (all the classes with the "Fixed" suffix) to make it work. But that also allowed us to upgrade to latest Jetty version 12.

Which is a good thing, because the Sparkjava project, which we used for delivering cached clip contents, also seems abandoned. So we phased that out as well and re-implemented using Jetty directly.

For this to work, we also had to bump the Java version. 17 would have done the job, but as 21 (LTS) was out just some days ago, we decided to switch to that.

Along with this we had to upgrade our OCI Base Image. We've choosen the Amazon Corretto Alpine image as new base. Amazon Corretto is well maintained, and Alpine is sufficient for our needs.